# api/openapi.yaml
openapi: "3.1.0"
info:
  title: go-auth API
  version: 0.2.0
  description: Reusable auth endpoints — mountable handlers for Go APIs. JWT + bcrypt + refresh tokens.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /auth/refresh:
    post:
      summary: Exchange refresh token for new token pair
      operationId: refreshToken
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid, expired, or revoked refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    AuthResponse:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token:
          type: string
          description: Short-lived JWT (15 min)
        refresh_token:
          type: string
          description: Long-lived opaque token (30 days)
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "email_taken"
            message:
              type: string
              description: Human-readable error message
              example: "A user with this email already exists"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT